moved pytorch models into agents, added passing of discrete actions model chose back ot model for loggin, and adding hidden layer sweep for feature analysis
