+++ models/pytorch/pytorch_models/fworld/config/config_inputs.yaml
