+++ models/pytorch/pytorch_models/fworld/config/run_config.yaml
@@ -2,7 +2,8 @@ name: "onpolicysupervised"
 train_runs: 10000
 max_runs: 10001
 description: "a simple model using inputs and outputs from fworld in a bp network"
-hidden_size: 125
+hidden_size:
+  - 125
 infer_offpolicy: False
 max_epochs: 10
 log_remote: False
