+++ models/pytorch/pytorch_models/fworld/config/run_config_offline.yaml
@@ -2,7 +2,11 @@ name: "onpolicy-trainsamples"
 train_runs: 10000
 max_runs: 10001
 description: "a simple model using inputs and outputs from fworld for ablations on their impact on the network. Explicitely examining ability to overfit"
-hidden_size: 125
+hidden_size:
+  - 10
+  - 125
+  - 500
+  - 1000
 infer_offpolicy: False
 max_epochs: 10
 datapath: "fworld_onpolicysupervised-2ets9y8o.pkl"
